+++ game.js
@@ -5,6 +5,8 @@ import { showOptionsP4, hideOptionsP4 } from "./optionsP4.js";
 import { showWeaponsOptions, hideWeaponOptions } from "./weaponOptions.js";
 import { addShovel, addFamiliarCloth } from "./items.js";
 import { showSlide, hideSlide } from "./slideshow.js";
+import { showCemetery, hideCemetery } from "./cemetery.js";
+import { showWeapon } from "./weapon.js";
 
 let player_name = "";
 let items = new Set();
@@ -214,23 +216,6 @@ const changeScene = (p) => {
   }
   console.log(`returning ${p}`);
   return p;
-};
-
-const showCemetery = (text) => {
-  $("#cemetery").removeClass("d-none");
-  $("#charpd").addClass("d-none");
-  $("#gameplay_text").html(text);
-  $("#gameplay_text").removeClass("d-none");
-  hideOptionsP2();
-};
-
-const hideCemetery = () => {
-  $("#cemetery").addClass("d-none");
-  $("#charpd").removeClass("d-none");
-};
-
-const showWeapon = (weapon) => {
-  $(`#${weapon}`).removeClass("d-none");
 };
 
 const gameScene = (p) => {
