+++ weapon.js
@@ -1,0 +1,5 @@
+const showWeapon = (weapon) => {
+  $(`#${weapon}`).removeClass("d-none");
+};
+
+export { showWeapon };
